pyproject.toml
[project]
name = "proj"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "flask>=3.1.3",
    "pandas>=2.3.3",
    "plotly>=6.6.0",
    "pymupdf>=1.27.2",
    "pyserial>=3.5",
    "requests>=2.32.5",
    "streamlit>=1.55.0",
    "streamlit-autorefresh>=1.0.1",
]
